[meta/api] rename MetaClientOnKV to SchemaApiImpl<KV:KVApi>: impl MetaApi on any KVApi impl
- fix: #5191
