+++ common/meta/api/Cargo.toml
@@ -17,8 +17,15 @@ common-base = { path = "../../base" }
 common-datavalues = { path = "../../datavalues" }
 common-exception = { path = "../../exception" }
 common-meta-types = { path = "../types" }
+common-proto-conv = { path = "../../proto-conv" }
+common-protos = { path = "../../protos" }
 common-tracing = { path = "../../tracing" }
 
+anyerror = "0.1.6"
 anyhow = "1.0.56"
 async-trait = "0.1.53"
 maplit = "1.0.2"
+serde = { version = "1.0.136", features = ["derive"] }
+serde_json = "1.0.79"
+thiserror = "1.0.30"
+tonic = { version = "=0.6.2", features = ["transport", "codegen", "prost", "tls-roots", "tls"] }
