+++ common/meta/api/src/lib.rs
@@ -16,12 +16,20 @@
 extern crate common_meta_types;
 
 mod kv_api;
+mod kv_api_key;
 mod kv_api_test_suite;
 mod meta_api;
 mod meta_api_test_suite;
+mod schema_api_impl;
+mod schema_api_keys;
 
 pub use kv_api::KVApi;
 pub use kv_api::KVApiBuilder;
+pub use kv_api_key::KVApiKey;
+pub use kv_api_key::KVApiKeyError;
 pub use kv_api_test_suite::KVApiTestSuite;
 pub use meta_api::MetaApi;
 pub use meta_api_test_suite::MetaApiTestSuite;
+pub use schema_api_impl::SchemaApiImpl;
+pub use schema_api_keys::DatabaseIdGen;
+pub use schema_api_keys::TableIdGen;
